==> 00_manage_packages.Rmd <==
---
title: "Manage packages"
author: "Taebin Kim"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

Install and load the package `pacman` for package management. Required 
packages will be added as we go along.

```{r setup, include=FALSE}
if (!require("pacman")) {
  install.packages("pacman")
  library(pacman)
} 
p_load(devtools,
       tidyr,
       tidyverse,
       dplyr,
       stringr,
       ggplot2,
       readxl,
       naniar,
       GGally,
       ggcorrplot,
       FactoMineR,
       factoextra,
       reshape2,
       visdat,
       ggfortify,
       fastDummies)
```
