==> 02_EDA.Rmd <==
---
title: "EDA"
author: "SooHyun Kim, Kyungjin Sohn"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

```{r, warning=FALSE, message=FALSE, results='hide'}
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

source_rmd("00_manage_packages.Rmd")
```

## 1. Demographics data
```{r read1}
load("./data/demographics.RData")
```

### A. Missing Data
* Out of 130 observations 45 rows have at least one missing input.
```{r miss_percentage1}
nrow(demographics_block[complete.cases(demographics_block),])/nrow(demographics_block)
nrow(demographics_block) - nrow(demographics_block[complete.cases(demographics_block),])
```

```{r heatmap1, fig.align='center', fig.width=9, fig.height=7}
vis_miss(demographics_block[,2:7]) + 
     theme(axis.text.x = element_text(angle = 75)) +
     labs(title = "Heatmap of Missing Data in the Demographics")
```

To be specific, 29 observations for Own puberty level, 21 observations for Parent's puberty level, and 6 observations for Average parent education variable are missing.
```{r miss_count1}
sapply(demographics_block, function(x) sum(is.na(x)))
```

One subject (`cnt_277_bl`) has missing values for all three variables.
```{r demo_na3}
allNA <- which(apply(demographics_block, MARGIN = 1, function(x) sum(is.na(x))) == 3)
demographics_block[allNA,]
```

Additionally, there are 9 subjects who have two missing values among the three columns. Of these 9 subjects, 6 have two missing values in the two puberty level columns, 2 have two missing values in both the Parental puberty level and Parental education average columns, and one have two missing values in both the own puberty level and the Parental education average.
```{r demo_na2}
twoNA <- which(apply(demographics_block, MARGIN = 1, function(x) sum(is.na(x))) == 2)
demographics_block[twoNA,]
```

#### Exploring the Missing Data Points

```{r explore_demo1, warning=FALSE, message=FALSE, fig.align='center'}
gg_miss_upset(demographics_block[,2:7], nsets = n_var_miss(demographics_block))
```

```{r explore_demo2, fig.align='center'}
# using  geom_miss_point()
ggplot(demographics_block,
       aes(x = PubertyTannerSelf,
           y = PubertyTannerParent)) +
 geom_miss_point()

# using  geom_miss_point()
ggplot(demographics_block,
       aes(x = PubertyTannerSelf,
           y = ParentalEducationAvg)) +
 geom_miss_point()

# using  geom_miss_point()
ggplot(demographics_block,
       aes(x = PubertyTannerSelf,
           y = ParentalEducationAvg)) +
 geom_miss_point()
```

If we remove observations with at least one missing value, there are 85 subjects remaining.
```{r explore_demo3}
nrow(demographics_block %>% drop_na(.))
```

### B. Summary of variables
```{r summary1}
summary(demographics_block)
```

```{r}
# save without missing data
demographics_block2 <- demographics_block %>% na.omit()
# save(demographics_block2, file = "./data/demographics2.RData")
```

### C. Associations among variables

* There are more male subjects than female.
* The Age, Own puberty level, and Parent's puberty level variables are fairly spread out.
* The Average parent education level variable is left-skewed, with most values ranging between 6 to 10.
* Among the dataset, females have a higher average own puberty level than males.
* There seems to be a strong positive correlation between Age in years and Parent's puberty level.
* Additionally, there is a high correlation between Own puberty level and Parent's puberty level.

```{r corr1, warning=FALSE, message=FALSE, fig.align='center', fig.width=9, fig.height=8}
ggpairs(demographics_block2, columns = c(2, 4:7), 
        title = "Scatterplot matrix (pairs plots) of variables")
```


## 2. Cortisol data
```{r read_2}
load("./data/cortisol.RData")
```

### A. Missing Data
* No missing data for block 1
```{r miss_percentage2}
nrow(cortisol_wide[complete.cases(cortisol_wide),])/nrow(cortisol_wide)
```

### B. Summary of variables
```{r summary2}
summary(cortisol_long)
summary(cortisol_wide)
```

### C. Associations among variables

* All regions are significantly correlated. (all over $0.3$)
* High correlation between `M_S1`,`M_S2`, and `M_S3` with correlations above $0.65$.
* High correlation between `M_S3`and `M_S4 `with a correlation $0.7$.
* High correlation between `M_S4`,`M_S5`, and `M_S6` with correlations higher than $0.75$.

```{r corr2, warning=FALSE, message=FALSE, fig.align='center', fig.width=9}
ggpairs(cortisol_wide, columns = c(2:7), 
        title = "Scatterplot matrix (pairs plots) of variables")
```

```{r boxplot2, warning=FALSE, fig.align='center'}
ggplot(cortisol_long, aes(x = as.factor(timepoint), y = cortisol)) +
  geom_boxplot()+
  ylim(0,.5)+
  labs(x = "timepoint")
```

Does not seem to be a clear shape in the cortisol over time.


## 3. fMRI data

```{r read_3}
load("./data/fMRI.RData")
```

### A. Missing Data
* No missing data for block 2
```{r miss_percentage3}
nrow(fmri_wide[complete.cases(fmri_wide),])/nrow(fmri_wide)
```

### B. Summary of variables
```{r summary3}
summary(fmri_wide)
```

### C. Associations among variables

* The corresponding left and right regions are positively correlated.
* Both the left and right sides of the `vmpfc` are highly correlated with each other. However, the right side is also correlated with the right side of the `dlpfc` and the `hippo`. On the other hand, the left side of the `vmpfc` is only correlated with the right side of the `hippo`.
* The `hippo_R_stor` is positively correlated with `amyg` on either side, while the `hippo_L_stor` is highly correlated with both `hippo` and `dlpfc` on either side.

```{r scatter2, warning=FALSE, message=FALSE, fig.align='center', fig.width=11, fig.height=9}
ggpairs(fmri_wide, columns = c(2:9), 
        title = "Scatterplot matrix (pairs plots) of variables")
```


## 4. Psych Profile data

```{r read_4}
load("./data/pysch.RData")
```

### A. Missing Data
* Out of 130 observations 8 rows have at least one missing input.
```{r miss_percentage4}
nrow(psych_block[complete.cases(psych_block),])/nrow(psych_block)
nrow(psych_block) - nrow(psych_block[complete.cases(psych_block),])
```

```{r heatmap4, fig.align='center', fig.width=9, fig.height=7}
vis_miss(psych_block) + 
     theme(axis.text.x = element_text(angle = 75)) +
     labs(title = "Heatmap of Missing Data in Block 3. Psych Profile")
```

```{r miss_count4}
sapply(psych_block, function(x) sum(is.na(x)))
```

From the plot above, we can see that 4% of the variable `STAIPost`-Pre are missing, which corresponds to 5 out of 130 observations. We can also check that `STAITrait` has 4 missing values. One observation has missing values for both variables.

#### Exploring the Missing Data Points
```{r explore_psych1, fig.align='center'}
gg_miss_upset(psych_block[,2:6], nsets = n_var_miss(psych_block))
```

```{r explore_psych2, fig.align='center'}
# using  geom_miss_point()
ggplot(psych_block,
       aes(x = STAITrait,
           y = `STAIPost-Pre`)) +
 geom_miss_point()

# using  geom_miss_point()
ggplot(psych_block,
       aes(x = CogDisScore,
           y = `STAIPost-Pre`)) +
 geom_miss_point()

# using  geom_miss_point()
ggplot(psych_block,
       aes(x = Anhedonia,
           y = `STAIPost-Pre`)) +
 geom_miss_point()

# using  geom_miss_point()
ggplot(psych_block,
       aes(x = CsopsNegative,
           y = `STAIPost-Pre`)) +
 geom_miss_point()
```

If we remove observations with at least one missing value, there are 122 subjects remaining.
```{r explore_psych3}
nrow(psych_block %>% drop_na(.))
psych_block2 <- psych_block %>%  drop_na(.)
```


```{r}
# save without missing data
psych_block2 <- psych_block %>%  na.omit()
# save(psych_block2, file = "./data/pysch2.RData")
```


### B. Summary of variables

```{r summary4}
summary(psych_block2)
```

### C. Associations among variables

* All the variables have a positive correlation except `STAIPost-Pre`.

```{r corr4, warning=FALSE, message=FALSE, fig.align='center', fig.width=9, fig.height=7}
ggpairs(psych_block2, columns = c(2:6), 
        title = "Scatterplot matrix (pairs plots) of variables")
```


## 5. Working memory data

```{r read_5}
load("./data/working.RData")
```

### A. Missing Data
* No missing data for block 6 Working memory.
```{r miss_percentage5}
nrow(working_wide[complete.cases(working_wide),])/nrow(working_wide)
nrow(working_wide) - nrow(working_wide[complete.cases(working_wide),])
```

### B. Summary of variables
```{r summary5}
summary(working_wide)
```

### C. Associations among variables

* Usually post-working memory has a positive correlation with each other, and a similar trend is found with pre-working memory
* We find some post- and pre- working memory to be negatively correlated.

```{r b5_scatter, warning=FALSE, message=FALSE, fig.align='center', fig.width=30, fig.height=22}
pairs_plot <- ggpairs(working_wide, columns = c(2:17), 
        title = "Scatterplot matrix (pairs plots) of variables")

pairs_plot
```


## 6. Autonomic data

```{r read_6}
load("./data/autonomic.RData")
```

### A. Missing Data
* Out of 174 observations 137 rows have at least one missing input.
```{r miss_percentage6}
nrow(autonomic_block[complete.cases(autonomic_block),])/nrow(autonomic_block)
nrow(autonomic_block) - nrow(autonomic_block[complete.cases(autonomic_block),])
```

```{r heatmap6, fig.align='center', fig.width=9, fig.height=7}
vis_miss(autonomic_block) + 
     theme(axis.text.x = element_text(angle = 75)) +
     labs(title = "Heatmap of Missing Data in the HRV and SAA")
```

Missing data points is a significant issue in the Autonomic block. Missing data appears in $47.3%$ of the data. The `SAA` variables seem to have a high number of missing values with over $50%$ for all `AA` variables. For the `HRV_MIST` variables have a missing rate all over $40%$

### Exploring the Patterns of Missing Data
```{r explore6, fig.align='center'}
gg_miss_upset(autonomic_block[,2:10], nsets = n_var_miss(autonomic_block))
```

53 observations have all missing inputs. We will drop this block from analysis due to missing values issue.

### B. Summary of variables
```{r summary6}
summary(autonomic_block)
```

### C. Associations among variables

* The 3 measurements of Heart Rate Variability (HRV) is highly correlated with each other.
* All 6 measurements of SAA are highly correlated with each other.

```{r scatter6, warning=FALSE, message=FALSE, fig.align='center', fig.width=10, fig.height=8}
ggpairs(autonomic_block, columns = c(2:10), 
        title = "Scatterplot matrix (pairs plots) of variables")
```
