==> 03_PCA.Rmd <==
---
title: 'PCA: Principal Component Analysis'
author: "Amy Song"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: html_document
---

```{r, message = F}
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

source_rmd("00_manage_packages.Rmd")
```
## Loading Data
```{r}
# Cortisol Data
load('data/cortisol.RData')
# fMRI Data
load('data/fmri.Rdata')
# Working Memory Data
load('data/working.Rdata')
# demographics Data
load('data/demographics.Rdata')
# psych data
load('data/pysch.Rdata')
```

# Background: PCA and Scree Plot

Principal Components Analysis is used to find a lower dimensional structure from the high dimensional data. For data with correlated variables, principal components allow us to summarize the set of variables with a smaller number of representative variables that collectively explain most of the variability in the original set [2]. 

For each of the data blocks given, it is in the form of a matrix, $X \in \mathbb{R}^{n\times p}$, where $n$ is the number of observations and $p$ is the number of features. When $p$ is large, a lower-dimensional representation of the data matrix may be enough to capture a lot of information from the original data. "PCA finds a small number of dimensions that are interesting as possible, where the concept of *interesting* is measured by the amount that the observations vary along each dimension. Each of the dimensions found by PCA is a linear combination of the $p$ features" [2]. 

To perform PCA, the variables in the data are centered to have a mean of zero. Depending on the presence of multiple measure units in the data, the data blocks are further scaled to have a standard deviation of 1. Among the data blocks used for PCA, data blocks on demographics and psychology profiles are scaled for these two data blocks containing variables with different measurement units. On the other hand, the data blocks related to cortisol measure, fMRI signals, and working memory are not scaled. The data are scaled for PCA because "the results obtained when we perform PCA will also depend on whether the variables have been individually scaled" [2]. 

In PCA, the proportion of variance explained is used to measure how much the selected principal components can explain the variance in the data. The scree plot visualizes the proportion of variance explained by each principal component. One chooses the smallest number of principal components that are required to explain a sizable amount of variation in the data [2]. For this project, the smallest number of principal components that explained 95% of the variation in the data is chosen for each data block. 

*Figure 7* displays the scree plot of PCA performed on the cortisol data block. The bars on the plot depict the percentage of variance explained by each of the principal components in the cortisol data block. The gray line depicts the cumulative percentage of variance explained by the principal components. As shown in the scree plot, the first three principal components explain about 95.5% of the variation in the data. Using this, it is decided to use 3 principal components for this data block. 

In addition to the proportion of variance explained by each principal component, loading vectors and scores can also be visualized. The loading vector, $\phi_i = [\phi_{1, i},\phi_{2, i}, \cdots \phi_{p, i}], i = 1, \cdots, p$, for each principal components defines a direction in feature space along with the data vary the most. The scores, $z_{1, i}, \cdots z_{n, i}, i = 1, \cdots, p$ for principal component $i$ are the values when the data with n observations are projected onto the direction of loading vector. The bi-plot visualizes the loading vectors and scores for two selected principal components. *Figure 8* displays the bi-plot of PCA on the cortisol data block. The points on the plot depict the score of each observation from the first and second principal components, and the red arrows are the loading vector for each variable. 


# PCA and Scree-Plot on data blocks
Using correlation matrix of each data block to standardize 
The number of components selected for each PCA of data block if 95% of the data is explained. 

## Cortisol (block 1)
Checking the dimension of the data block and whether there is `Subject` in the column. This column will be not be considered in the PCA. 
```{r}
dim(cortisol_wide)
names(cortisol_wide)
```

```{r, warning = F}
# PCA of the block: the matrix is scaled
cortisol_pca<- prcomp(cortisol_wide[,2:7], scale = F)
# Scree plot of the PCA
cumulated.var <- get_eig(cortisol_pca)$cumulative.variance.percent
cumulated.var.labs <- paste(round(cumulated.var,1),'%')
cumulated.var.labs[1] <- NA
factoextra::fviz_eig(cortisol_pca, addlabels = T, title = "Scree Plot for Cortisol")  + 
  geom_line(aes(y = cumulated.var), color = "lightslategrey") + 
  geom_point(aes(y = cumulated.var), color = "lightslategrey") +
  geom_text(aes(label = cumulated.var.labs, y = cumulated.var + 5), color = "lightslategrey")
ggsave("./plots/cortisol_pca.png", bg = "white")
```
```{r}
# Summary of the PCA
summary(cortisol_pca)
```
* As seen in the scree plot and the summary of the PCA, *3 Components* explains 99% of the variances in the data. The following is the biplot, which plots both scores and loading values of the first 2 principal components. The points on the plot indicate the scores of each observations and the blue errors indicate the loading vectors of the variables. For instance, the loading value for `M_S3` on the first PC is 0.2795931 and 0.3229215 for second PC, so the vector is centered at (0.2795931, 0.3229215)

```{r}
# Bi-Plot of the PCA: both scores and loading
autoplot(cortisol_pca, label = FALSE, loadings.label = TRUE, loadings.label.repel = TRUE, title = "Bi-Plot of Cortisol") + 
  geom_vline(xintercept = 0, color = "gray28", lwd=0.25) + 
  geom_hline(yintercept = 0, color = "gray28", lwd=0.25) 
ggsave("./plots/cortisol_bi-plot.png", bg = "white")
# loading values of PCA
cortisol_pca$rotation
```

```{r}
# Bar plot of the Contributions of each variables in the selected PC
factoextra::fviz_contrib(cortisol_pca,
                 choice = "var",
                 axes = 3)
```

## fMRI Block
Checking the dimension of the data block and whether there is `Subject` in the column. This column will be not be considered in the PCA. 
```{r}
dim(fmri_wide)
names(fmri_wide)
```

```{r, warning = F}
# PCA of the block: the matrix is centered
fmri_pca <- prcomp(fmri_wide[,2:9])
# Scree plot of the PCA
cumulated.var <- get_eig(fmri_pca)$cumulative.variance.percent
cumulated.var.labs <- paste(round(cumulated.var,1),'%')
cumulated.var.labs[1] <- NA
factoextra::fviz_eig(fmri_pca, addlabels = T, title = "Scree Plot for fMRI")  + 
  geom_line(aes(y = cumulated.var), color = "lightslategrey") + 
  geom_point(aes(y = cumulated.var), color = "lightslategrey") +
  geom_text(aes(label = cumulated.var.labs, y = cumulated.var + 5), color = "lightslategrey")
ggsave("./plots/fmri_pca.png", bg = "white")
```
```{r}
# Summary of the PCA
summary(fmri_pca)
```
* As seen in the scree plot and the summary of the PCA, *5 Components* explains 95% of the variances in the data. The following is the biplot, which plots both scores and loading values of the first 2 principal components. The points on the plot indicate the scores of each observations and the blue errors indicate the loading vectors of the variables. For instance, the loading value for `amyg_R_stor` on the first PC is -0.4315 and 0.0300 for second PC, so the vector is centered at (-0.4315, 0.0300)

```{r}
# Bi-Plot of the PCA: both scores and loading
ggplot2::autoplot(fmri_pca, label = FALSE, loadings.label = TRUE, loadings.label.repel = T) + 
  geom_vline(xintercept = 0, color = "gray28", lwd=0.25) + 
  geom_hline(yintercept = 0, color = "gray28", lwd=0.25) 
ggsave("./plots/fmri_bi-plot.png", bg = "white")
# loading values of PCA
fmri_pca$rotation
```
```{r}
# Bar plot of the Contributions of each variables in the selected PC
factoextra::fviz_contrib(fmri_pca,
                 choice = "var",
                 axes = 5)
```

## Working Memory 

Checking the dimension of the data block and whether there is `Subject` in the column. 
```{r}
dim(working_wide)
names(working_wide)
```

```{r, warning = F}
# PCA of the block: the matrix is centered
#wm_corr <- cor(wm_df)
wm_pca <- prcomp(working_wide[,2:17])
# Scree plot of the PCA
cumulated.var <- get_eig(wm_pca)$cumulative.variance.percent[1:10]
cumulated.var.labs <- paste(round(cumulated.var,1),'%')
cumulated.var.labs[1] <- NA
factoextra::fviz_eig(wm_pca, addlabels = T, title = "Scree Plot for Working Memory")  + 
  geom_line(aes(y = cumulated.var), color = "lightslategrey") + 
  geom_point(aes(y = cumulated.var), color = "lightslategrey") +
  geom_text(aes(label = cumulated.var.labs, y = cumulated.var + 5), color = "lightslategrey")
ggsave("./plots/wm_pca.png", bg = "white")
```
```{r}
# Summary of the PCA
summary(wm_pca)
```

* As seen in the scree plot and the summary of the PCA, *9 Components* explains 95% of the variances in the data. The following is the biplot, which plots both scores and loading values of the first 2 principal components. The points on the plot indicate the scores of each observations and the blue errors indicate the loading vectors of the variables. For instance, the loading value for `amyg_L_stor_PostStress` on the first PC is 0.039938048 and -0.174290802 for second PC, so the vector is centered at (0.039938048, -0.174290802)

```{r}
# Bi-Plot of the PCA: both scores and loading
ggplot2::autoplot(wm_pca, label = FALSE, loadings.label = TRUE, loadings.label.repel = T) + 
  geom_vline(xintercept = 0, color = "gray28", lwd=0.25) + 
  geom_hline(yintercept = 0, color = "gray28", lwd=0.25) 
ggsave("./plots/wm_bi-plot.png", bg = "white")
# loading values of PCA
wm_pca$rotation[,c(1,2,9)]
```

## Demographics 

```{r}
dim(demographics_block)
names(demographics_block)
```

```{r, warning=F}
# PCA of the block: the matrix is centered and scaled
demo_pca <- demographics_block %>% mutate(Sex = as.numeric(Sex)) %>% select(-c(SubjectID,AgeInYears)) %>% drop_na(.) %>% prcomp(scale = T)
# Scree plot of PCA with cumulated sum of percentage of variance explained
cumulated.var <- get_eig(demo_pca)$cumulative.variance.percent
cumulated.var.labs <- paste(round(cumulated.var,1),'%')
cumulated.var.labs[1] <- NA
factoextra::fviz_eig(demo_pca, addlabels = T, title = "Scree Plot for Demographics")  + 
  geom_line(aes(y = cumulated.var), color = "lightslategrey") + 
  geom_point(aes(y = cumulated.var), color = "lightslategrey") +
  geom_text(aes(label = cumulated.var.labs, y = cumulated.var + 5), color = "lightslategrey")
ggsave("./plots/demo_pca.png", bg = "white")
```
```{r}
summary(demo_pca)
```
* As seen in the scree plot and the summary of the PCA, *4 Components* explains at least 95% of the variances in the data. The following is the biplot, which plots both scores and loading values of the first 2 principal components.

```{r}
# Bi-Plot of the PCA: both scores and loading
ggplot2::autoplot(demo_pca, label = FALSE, loadings.label = TRUE, loadings.label.repel = T, data = demographics_block %>% select(-c(SubjectID,AgeInYears)) %>% drop_na(.), colour = 'Sex')  + 
  geom_vline(xintercept = 0, color = "gray28", lwd=0.25) + 
  geom_hline(yintercept = 0, color = "gray28", lwd=0.25) 
ggsave("./plots/demo_bi-plot.png", bg = "white")
# loading values of PCA
demo_pca$rotation
# color bi-plot
```

```{r}
# Bar plot of the Contributions of each variables in the selected PC
factoextra::fviz_contrib(demo_pca,
                 choice = "var",
                 axes = 4)
```

## Psych Profile
drop the rows with missing values
```{r}
dim(psych_block)
names(psych_block)
```

```{r, warning=F}
psych_pca <- psych_block %>% select(-SubjectID) %>% drop_na(.) %>% prcomp(scale = T)
# Scree plot of PCA with cumulated sum of percentage of variance explained
cumulated.var <- get_eig(psych_pca)$cumulative.variance.percent
cumulated.var.labs <- paste(round(cumulated.var,1),'%')
cumulated.var.labs[1] <- NA
factoextra::fviz_eig(psych_pca, addlabels = T, title = "Scree Plot for Psych Profile")  + 
  geom_line(aes(y = cumulated.var), color = "lightslategrey") + 
  geom_point(aes(y = cumulated.var), color = "lightslategrey") +
  geom_text(aes(label = cumulated.var.labs, y = cumulated.var + 5), color = "lightslategrey")
ggsave("./plots/psych_pca.png", bg = "white")
```

```{r}
summary(psych_pca)
```

* As seen in the scree plot and the summary of the PCA, *4 Components* explains 95% of the variances in the data. The following is the biplot, which plots both scores and loading values of the first 2 principal components.

```{r}
# Bi-Plot of the PCA: both scores and loading
ggplot2::autoplot(psych_pca, label = FALSE, loadings.label = TRUE, loadings.label.repel = T) + 
  geom_vline(xintercept = 0, color = "gray28", lwd=0.25) + 
  geom_hline(yintercept = 0, color = "gray28", lwd=0.25) 
ggsave("./plots/psych_bi-plot.png", bg = "white")
# loading values of PCA
psych_pca$rotation[,c(1,2,4)]
```

```{r}
# Bar plot of the Contributions of each variables in the selected PC
factoextra::fviz_contrib(psych_pca,
                 choice = "var",
                 axes = 4)
```
