==> 04_AJIVE.Rmd <==
---
title: "AJIVE Analysis"
author: "Taebin Kim & Thomas H. Keefe"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document: default
  pdf_document: default
---

```{r, warning=FALSE, message=FALSE, results='hide'}
source_rmd = function(file, ...) {
  tmp_file = tempfile(fileext = ".R")
  on.exit(unlink(tmp_file), add = TRUE)
  knitr::purl(file, output = tmp_file)
  source(file = tmp_file, ...)
}

source_rmd("00_manage_packages.Rmd")
```
## Setup

Install the `r_jive` package.
```{r}
devtools::install_github("taebinkim7/r_jive")
library(ajive)
```

Load preprocessed data blocks.
```{r}
load("data/cortisol.RData")
load("data/fmri.RData")
load("data/working.RData")
load("data/demographics2.RData")
```


Set signal ranks based on the PC scree plots.
```{r}
cortisol_signal_rank <- 3
fmri_signal_rank <- 5
working_signal_rank <- 9
demographics_signal_rank <- 4
```

Redefine data blocks for the AJIVE analyses.
```{r}
cortisol = cortisol_wide
fmri = fmri_wide
working = working_wide
demographics = demographics_block2 %>%
  fastDummies::dummy_cols("Sex") %>%
  select(-c(Sex, Sex_F, AgeInYears))
```

Define a function for fitting an AJIVE model.

* Usage
```{r, eval=FALSE}
fit_ajive(data_blocks, signal_ranks, n_wedin_samples = 100, n_rand_dir_samples = 100)
```

* Arguments

`data_blocks`: List of data blocks for AJIVE analysis

`signal_ranks`: Predetermined signal ranks for the data blocks. The length should be the same as the number of data blocks.

`n_wedin_samples`: Number of samples for obtaining the Wedin bound. Default is 100.

`n_rand_dir_samples`: Number of random direction samples. Default is 100.

```{r}
merge_data <- function(data_blocks){
  merged_data_block <- Reduce(
      function (...){
          merge(..., all = FALSE)
      }, 
      data_blocks
  )
  common <- merged_data_block$SubjectID
  
  inter_data_blocks <- sapply(
      data_blocks,
      function(x){
          as.matrix(x[x$SubjectID %in% common, ] %>% select(-1))
      }
  )
  return(inter_data_blocks)
}
```


```{r}
fit_ajive <- function(merged_data_blocks,
                      signal_ranks,
                      n_wedin_samples = 100,
                      n_rand_dir_samples = 100) {
  
  ajive <- ajive::ajive(merged_data_blocks,
                        signal_ranks,
                        n_wedin_samples = n_wedin_samples,
                        n_rand_dir_samples = n_rand_dir_samples)
  return(ajive)
}
```

## AJIVE Analyses

### Cortisol and FMRI

* Q1. Investigate relationship between stress response profile (cortisol timeseries block) and brain region response (FMRI block) during stress test.


```{r}
merged_cortisol_fmri <- merge_data(list(cortisol, fmri))
```

```{r}
ajive_cortisol_fmri <- ajive(
  merged_cortisol_fmri,
  c(cortisol_signal_rank, fmri_signal_rank),
)
```

#### Joint rank
```{r}
cat("Joint rank:", ajive_cortisol_fmri$joint_rank)
```

#### Common Normalized Score
```{r}
cns <- ajive_cortisol_fmri$joint_scores
cns_df <- data.frame(cns=cns)
cns_df %>% ggplot(aes(x=cns)) + geom_density() + geom_rug() + ggtitle("Common normalized scores") + theme_bw()
```

```{r}
scores_cortisol <- as_tibble(x=cbind(ajive_cortisol_fmri$block_decomps[[1]][['joint']][['u']],
                                   ajive_cortisol_fmri$block_decomps[[1]][['individual']][['u']])) %>% 
                    rename(joint=V1, indiv=V2)
                    
             
ggplot(scores_cortisol) + 
    geom_point(aes(x=indiv, y=joint)) + 
    geom_hline(yintercept=0) + 
    geom_vline(xintercept=0) +
    theme_bw() +
    ggtitle("The cortisol individual scores are correlated with the joint scores.")
```
```{r}
get_loadings <- function(scores, datablock){
  loadings <- t(t(scores) %*% datablock)  # Project scores onto datablock
  loadings <- loadings / sqrt(sum(loadings^2))  # Normalize the loadings vector
  loadings_df <- data.frame(loading=loadings) %>% rownames_to_column("variable")
  return(loadings_df)
}
```

```{r}
cort_joint_loads = get_loadings(ajive_cortisol_fmri$joint_scores, merged_cortisol_fmri[[1]])
cort_joint_loads$block = "cortisol"

fmri_joint_loads = get_loadings(ajive_cortisol_fmri$joint_scores, merged_cortisol_fmri[[2]])
fmri_joint_loads$block = "fMRI"

# Concatenate
cort_fmri_joint_loads = rbind(cort_joint_loads, fmri_joint_loads)
cort_fmri_joint_loads$variable = factor(cort_fmri_joint_loads$variable, levels=cort_fmri_joint_loads$variable)
cort_fmri_joint_loads
```

```{r}
cort_fmri_joint_loads %>% ggplot(aes(x=variable, y=loading)) + 
  geom_col(aes(fill=block)) + coord_flip() + theme_bw() +
  ggtitle("Cortisol and fMRI Joint Loadings")
```


### Cortisol and Working Memory

* Q2. Investigate relationship between stress response profile (cortisol timeseries block) and brain region response (working memory block) during working memory task.
```{r}
merged_cortisol_working <- merge_data(list(cortisol, working))
ajive_cortisol_working <- fit_ajive(
  merged_cortisol_working,
  c(cortisol_signal_rank, working_signal_rank),
)
```

```{r}
cns <- ajive_cortisol_working$joint_scores
cns_df <- data.frame(cns=cns)
cns_df %>% ggplot(aes(x=cns)) + geom_density() + geom_rug() + ggtitle("Common normalized scores") + theme_bw()
```

```{r}
cort_joint_loads = get_loadings(ajive_cortisol_working$joint_scores, merged_cortisol_working[[1]])
cort_joint_loads$block = "cortisol"

work_joint_loads = get_loadings(ajive_cortisol_working$joint_scores, merged_cortisol_working[[2]])
work_joint_loads$block = "working memory"

# Concatenate
cort_work_joint_loads = rbind(cort_joint_loads, work_joint_loads)
cort_work_joint_loads$variable = factor(cort_work_joint_loads$variable, levels=cort_work_joint_loads$variable)
cort_work_joint_loads

# Plot
cort_work_joint_loads %>% ggplot(aes(x=variable, y=loading)) + 
  geom_col(aes(fill=block)) + coord_flip() + theme_bw() + 
  ggtitle("Cortisol and Working Memory Joint Loadings")
```

#### Joint rank
```{r}
cat("Joint rank:", ajive_cortisol_working$joint_rank)
```



### Demographics as the third block

* Q3. Repeat Q1 and Q2 with demographics block as the third block.

```{r}
merged_cortisol_fmri_demographics <- merge_data(list(cortisol, fmri, demographics))
ajive_cortisol_fmri_demographics <- fit_ajive(
  merged_cortisol_fmri_demographics,
  c(cortisol_signal_rank, fmri_signal_rank, demographics_signal_rank)
)
```

#### Joint rank
```{r}
cat("Joint rank:", ajive_cortisol_fmri_demographics$joint_rank)
```


#### Common Normalized Score
```{r}
cns <- ajive_cortisol_fmri_demographics$joint_scores
cns_df <- data.frame(cns=cns)
cns_df$SexM <- factor(merged_cortisol_fmri_demographics[[3]][,5])
cns_df %>% ggplot(aes(x=cns)) + geom_density() + geom_rug(aes(color=SexM)) + ggtitle("Common normalized scores") + theme_bw() +
  geom_density(aes(group=SexM, fill=SexM, color=SexM), alpha=.3)
```

We see that the scores are bimodal, but not owing to sex.

#### Joint loadings
```{r}
cort_joint_loads = get_loadings(ajive_cortisol_fmri_demographics$joint_scores, merged_cortisol_fmri_demographics[[1]])
cort_joint_loads$block = "cortisol"

fmri_joint_loads = get_loadings(ajive_cortisol_fmri_demographics$joint_scores, merged_cortisol_fmri_demographics[[2]])
fmri_joint_loads$block = "fMRI"

demo_joint_loads = get_loadings(ajive_cortisol_fmri_demographics$joint_scores, merged_cortisol_fmri_demographics[[3]])
demo_joint_loads$block = "demo"

# Concatenate
cort_fmri_demo_joint_loads = rbind(cort_joint_loads, fmri_joint_loads, demo_joint_loads)
cort_fmri_demo_joint_loads$variable = factor(cort_fmri_demo_joint_loads$variable, levels=cort_fmri_demo_joint_loads$variable)
cort_fmri_demo_joint_loads

# Plot
cort_fmri_demo_joint_loads %>% ggplot(aes(x=variable, y=loading)) + 
  geom_col(aes(fill=block)) + coord_flip() + theme_bw() + 
  ggtitle("Cortisol, fMRI, and Demographics Joint Loadings")
```

### Cortisol, Working Memory, and Demographics

```{r}
merged_cortisol_working_demographics <- merge_data(list(cortisol, working, demographics))
ajive_cortisol_working_demographics <- fit_ajive(
  merged_cortisol_working_demographics,
  c(cortisol_signal_rank, working_signal_rank, demographics_signal_rank)
)
```

#### Joint rank
```{r}
cat("Joint rank:", ajive_cortisol_working_demographics$joint_rank)
```

#### Common Normalized Scores
```{r}
cns <- ajive_cortisol_working_demographics$joint_scores
cns_df <- data.frame(cns=cns)
cns_df$SexM <- factor(merged_cortisol_working_demographics[[3]][,5])
cns_df %>% ggplot(aes(x=cns)) + geom_density() + geom_rug(aes(color=SexM)) + ggtitle("Common normalized scores") + theme_bw() +
  geom_density(aes(group=SexM, fill=SexM, color=SexM), alpha=.3)
```


#### Joint loadings
```{r}
cort_joint_loads = get_loadings(ajive_cortisol_working_demographics$joint_scores, merged_cortisol_working_demographics[[1]])
cort_joint_loads$block = "cortisol"

work_joint_loads = get_loadings(ajive_cortisol_working_demographics$joint_scores, merged_cortisol_working_demographics[[2]])
work_joint_loads$block = "working memory"

demo_joint_loads = get_loadings(ajive_cortisol_working_demographics$joint_scores, merged_cortisol_working_demographics[[3]])
demo_joint_loads$block = "demo"

# Concatenate
cort_fmri_demo_joint_loads = rbind(cort_joint_loads, work_joint_loads, demo_joint_loads)
cort_fmri_demo_joint_loads$variable = factor(cort_fmri_demo_joint_loads$variable, levels=cort_fmri_demo_joint_loads$variable)

cort_fmri_demo_joint_loads %>% ggplot(aes(x=variable, y=loading)) + 
  geom_col(aes(fill=block)) + coord_flip() + theme_bw() +
  ggtitle("Cortisol, Working Memory, \nand Demographics Joint Loadings")
```
